Dialer: Add special code aliases
Android already has a mechanism for special codes that transforms
*#*#<code>#*#* into intents that can be caught by special intent
filters in application manifests.

However, sometimes, we have requirements to provide specific codes
to launch self-testing or factory-testing kinds of apps. Instead of
adding all those and associating them with intents, this patch adds
aliases that are treated the same way as the built-in special codes.

This works by configuring two arrays:
- special_code_aliases: The codes (such as *#123#, ##100#, etc, any will do).
- special_code_mapping: The converted codes, usually these would be (123, 100, etc)
                        but they are not necessarily the same numeric parts as
                        the aliases.

These should be configured by overriding the special codes per-device.

Change-Id: I01dc06db3210a6b775144fa8c057cd7d0e58d85c